cleanup: migrate core/auth
