+++ packages/core/src/auth/helpers.ts
