+++ packages/core/src/core/index.ts
@@ -1,8 +1,5 @@
-import { startServerAndOpenBrowser } from "./auth";
-import { Poll, Endpoints } from "../utils";
-import type { Manifest, Frameworks, UserInfo } from "../models";
+import type { Manifest, Frameworks } from "../models";
 import * as Communication from "./communication";
-import { PrismicSharedConfigManager } from "../filesystem/PrismicSharedConfig";
 import { Repositories } from "../models/Repositories";
 export * as Communication from "./communication";
 export interface Core {
@@ -62,57 +59,3 @@ export default function createCore({ cwd, base, manifest }: CoreParams): Core {
     },
   };
 }
-
-async function startAuth({
-  base,
-  url,
-  action,
-}: {
-  base: string;
-  url: string;
-  action: "signup" | "login";
-}): Promise<void> {
-  const { onLoginFail } = await startServerAndOpenBrowser(url, action, base);
-  try {
-    // We wait 3 minutes before timeout
-    await Poll.startPolling<UserInfo | null, UserInfo>(
-      () => Auth.validateSession(base),
-      (user): user is UserInfo => !!user,
-      3000,
-      60
-    );
-    return;
-  } catch (e) {
-    onLoginFail();
-  }
-}
-
-export const Auth = {
-  login: async (base: string): Promise<void> => {
-    const endpoints = Endpoints.buildEndpoints(base);
-    return startAuth({
-      base,
-      url: endpoints.Dashboard.cliLogin,
-      action: "login",
-    });
-  },
-  signup: async (base: string): Promise<void> => {
-    const endpoints = Endpoints.buildEndpoints(base);
-    return startAuth({
-      base,
-      url: endpoints.Dashboard.cliSignup,
-      action: "signup",
-    });
-  },
-  logout: (): void => PrismicSharedConfigManager.remove(),
-  validateSession: async (requiredBase: string): Promise<UserInfo | null> => {
-    const config = PrismicSharedConfigManager.get();
-
-    if (!config.cookies.length) return Promise.resolve(null); // default config, logged out.
-    if (requiredBase != config.base) return Promise.resolve(null); // not the same base so it doesn't count.
-
-    return Communication.validateSession(config.cookies, requiredBase).catch(
-      () => null
-    );
-  },
-};
